requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
django==3.2
django-filter==21.1
django-cors-headers 
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-extra-fields
gunicorn==20.0.4
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
packaging==21.3
pluggy==0.13.1
psycopg2-binary==2.9.6
pyparsing==3.0.7
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.20.0
toml==0.10.2
typing_extensions==4.1.1
urllib3==1.26.9
zipp==3.7.0
djoser
webcolors==1.11.1
Pillow==9.0.0